Create Digital clock using an Arduino and an OLED.ino
We use the RTClib library for interfacing with the DS3231 RTC module.
The Adafruit_SSD1306 library is used for the OLED display.
The current time is obtained from the RTC and displayed on the OLED screen.
The loop repeats every second to update the time.
Make sure to install the required libraries using the Arduino Library Manager. Adjust the code based on your specific RTC and OLED display configurations.